requirements.txt
beautifulsoup4==4.13.4
certifi==2025.1.31
charset-normalizer==3.4.1
frozendict==2.4.6
idna==3.10
multitasking==0.0.11
narwhals==1.35.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
peewee==3.17.9
platformdirs==4.3.7
plotly==6.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
six==1.17.0
soupsieve==2.6
ta==0.11.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
yfinance==0.2.55
